♻️ Reorganize auth feature layout
