--- packages/server/src/modules/auth.ts
@@ -1,24 +1,0 @@
-import crypto from 'crypto';
-
-export const createPasswordHash = async (password: string) => {
-    const salt = crypto.randomBytes(16).toString('hex');
-    const hash = crypto.pbkdf2Sync(password, salt, 1000, 64, 'sha512').toString('hex');
-    return `${salt}:${hash}`;
-};
-
-export const comparePassword = async (password: string, storedHash: string) => {
-    const [salt, hash] = storedHash.split(':');
-    const newHash = crypto.pbkdf2Sync(password, salt, 1000, 64, 'sha512').toString('hex');
-    return hash === newHash;
-};
-
-export const compareSharedSecret = (expected: string, received: string) => {
-    const expectedBuffer = Buffer.from(expected, 'utf8');
-    const receivedBuffer = Buffer.from(received, 'utf8');
-
-    if (expectedBuffer.length !== receivedBuffer.length) {
-        return false;
-    }
-
-    return crypto.timingSafeEqual(expectedBuffer, receivedBuffer);
-};
