+++ packages/server/src/routes/auth-pages.ts
@@ -1,6 +1,10 @@
 import { Router } from 'express';
+import {
+    createLoginPageHandler,
+    createLoginPageSubmitHandler,
+    createLogoutPageHandler,
+} from '../features/auth/http/pages.js';
 import type { AuthConfig } from '../modules/auth-mode.js';
-import { createLoginPageHandler, createLoginPageSubmitHandler, createLogoutPageHandler } from '../views/auth.js';
 
 export const createAuthPagesRouter = (authConfig: AuthConfig) =>
     Router()
